django/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
coverage==6.3.1
Django==4.0.2
djangorestframework==3.13.1
pysqlite3-binary==0.4.6
pytz==2021.3
sqlparse==0.4.2
djangorestframework-simplejwt